rtl_bt: firmware allowed any previously seen device to wake the deck
If a device was most recently been connected to the (LCD) deck (from
the device's point of view), even if its wake on bluetooth permission
had been set to "do not allow", even if the device had been removed
from the known devices list, the wobt implementation would allow it
to wake us.

The only way to disable these unwanted wake ups is to revert the
firmware to remove this trigger-happy wobt support.

Signed-off-by: Vivek Das Mohapatra <vivek@collabora.com>