Add BDF for board-id: 0x030a
The BDF fixes for the OLED deck wifi had not been upstreamed
and got dropped when upstream changed from board.bin to board-2.bin.

I believe this section indicates the correct board:

  00075670: 0000 0000 6b00 0000 6275 733d 7063 692c  ....k...bus=pci,
  00075680: 7665 6e64 6f72 3d31 3763 622c 6465 7669  vendor=17cb,devi
  00075690: 6365 3d31 3130 332c 7375 6273 7973 7465  ce=1103,subsyste
  000756a0: 6d2d 7665 6e64 6f72 3d31 3763 622c 7375  m-vendor=17cb,su
  000756b0: 6273 7973 7465 6d2d 6465 7669 6365 3d30  bsystem-device=0
  000756c0: 3130 382c 716d 692d 6368 6970 2d69 643d  108,qmi-chip-id=
  000756d0: 3138 2c71 6d69 2d62 6f61 7264 2d69 643d  18,qmi-board-id=
  000756e0: 3737 376d 0100 0000 fce9 0000 7f45 4c46  777m.........ELF

qm-board-id=777m <- This board ID is the one we care about.

The 777m blob is present in the previous version committed in this
repo _but_ its version had regressed upstream, so when we synced
linux-firmware we lost some fixes/improvements.

The version of the 777m section of board-2.bin is not obviously
encoded where we can see it - we need to chase the vendor for how
to check that.

This commit restores those fixes.

[from commit 916541c6 by @qingzong.ma]

Signed-off-by: Vivek Das Mohapatra <vivek@collabora.com>